fix typo in SearchDaoImpl class

+++ dao/src/main/java/com/getjavajob/training/timashovy/socialnetwork/dao/interfaces/SearchDao.java
@@ -5,6 +5,6 @@ import java.util.List;
 public interface SearchDao<T> {
 
     List<T> findResults(String searchQuery, int currentPage, int numOfRecords);
-    int findResultsAmount(String searchQuery);
+    Long findResultsAmount(String searchQuery);
 
 }
